requirements.txt
argparse 
configparser 
tensorflow 
Pillow
